+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.7.0-rc.4",
+  sdkVersion: "3.7.0-rc.5",
   genVersion: "2.616.1",
-  userAgent: "speakeasy-sdk/typescript 3.7.0-rc.4 2.616.1 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.7.0-rc.5 2.616.1 2.0 @orq-ai/node",
 } as const;